+++ client/src/bin/test_harness.rs
@@ -177,6 +177,7 @@ fn apply_spec(spec: &TestCharSpec, opt: &Opt, runtime: &Arc<Runtime>) -> Result<
             &opt.password,
             None,
             |_| false,
+            || None,
             &|_| {},
             |_| {},
             PathBuf::new(),